Move verbatim to VerbatimEnvironment in __init__.py
This also rewrites NoCharSubEnvironment using what verbatim does. This
is cleaner because it overwrites normalize instead of invoke, and more
classes will want to have custom invoke than custom normalize. This
means it is generally easier to inherit NoCharSubEnvironment.
